# docs/openapi.json
{
  "components": {
    "schemas": {
...
      "CancelOrderRequest": {
        "additionalProperties": false,
        "description": "Cancel an order before it is in progress.",
        "properties": {
          "reason": {
            "anyOf": [
              {
                "maxLength": 255,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Reason"
          }
        },
        "title": "CancelOrderRequest",
        "type": "object"
      },
...
      "CreateOrderRequest": {
        "additionalProperties": false,
        "description": "Create a gift-request order.",
        "properties": {
          "delivery_city": {
            "anyOf": [
              {
                "maxLength": 100,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "description": "Legacy city name; select delivery_city_id from GET /api/cities.",
            "title": "Delivery City"
          },
          "delivery_city_id": {
            "anyOf": [
              {
                "format": "uuid",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Delivery City Id"
          },
          "delivery_date": {
            "description": "Requested delivery date (<= 6 months out).",
            "format": "date",
            "title": "Delivery Date",
            "type": "string"
          },
          "description": {
            "anyOf": [
              {
                "maxLength": 2000,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Description"
          },
          "request_media_keys": {
            "description": "Confirmed request-photo keys (0\u20133).",
            "items": {
              "type": "string"
            },
            "maxItems": 3,
            "title": "Request Media Keys",
            "type": "array"
          }
        },
        "required": [
          "delivery_date"
        ],
        "title": "CreateOrderRequest",
        "type": "object"
      },
...
      "DisputeRequest": {
        "additionalProperties": false,
        "description": "Open a dispute on an order.",
        "properties": {
          "reason": {
            "maxLength": 1000,
            "minLength": 3,
            "title": "Reason",
            "type": "string"
          }
        },
        "required": [
          "reason"
        ],
        "title": "DisputeRequest",
        "type": "object"
      },
      "DisputeResponse": {
        "description": "A dispute's state.",
        "properties": {
          "id": {
            "title": "Id",
            "type": "string"
          },
          "order_id": {
            "title": "Order Id",
            "type": "string"
          },
          "reason": {
            "title": "Reason",
            "type": "string"
          },
          "resolution_note": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Resolution Note"
          },
          "status": {
            "title": "Status",
            "type": "string"
          }
        },
        "required": [
          "id",
          "order_id",
          "status",
          "reason"
        ],
        "title": "DisputeResponse",
        "type": "object"
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "OrderDetail": {
        "description": "A full order view for its participants.",
        "properties": {
          "assigned_at": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Assigned At"
          },
          "courier_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Courier Id"
          },
          "created_at": {
            "title": "Created At",
            "type": "string"
          },
          "current_actor_has_rated": {
            "description": "Whether the authenticated actor has rated this order.",
            "title": "Current Actor Has Rated",
            "type": "boolean"
          },
          "customer_id": {
            "title": "Customer Id",
            "type": "string"
          },
          "delivery_city": {
            "title": "Delivery City",
            "type": "string"
          },
          "delivery_city_id": {
            "format": "uuid",
            "title": "Delivery City Id",
            "type": "string"
          },
          "delivery_date": {
            "title": "Delivery Date",
            "type": "string"
          },
          "description": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Description"
          },
          "id": {
            "title": "Id",
            "type": "string"
          },
          "status": {
            "title": "Status",
            "type": "string"
          },
          "total_amount": {
            "title": "Total Amount",
            "type": "string"
          }
        },
        "required": [
          "id",
          "status",
          "customer_id",
          "courier_id",
          "delivery_city",
          "delivery_city_id",
          "delivery_date",
          "description",
          "total_amount",
          "assigned_at",
          "created_at",
          "current_actor_has_rated"
        ],
        "title": "OrderDetail",
        "type": "object"
      },
      "OrderListResponse": {
        "description": "A keyset page of order summaries.",
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/OrderSummary"
            },
            "title": "Items",
            "type": "array"
          },
          "next_cursor": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Next Cursor"
          }
        },
        "required": [
          "items"
        ],
        "title": "OrderListResponse",
        "type": "object"
      },
...
      "OrderStatus": {
        "description": "Order lifecycle status (see app/services/order_state.py).",
        "enum": [
          "NEW",
          "ASSIGNED",
          "WAITING_PAYMENT",
          "IN_PROGRESS",
          "DELIVERED",
          "COMPLETED",
          "CANCELLED",
          "DISPUTED",
          "REFUNDED"
        ],
        "title": "OrderStatus",
        "type": "string"
      },
      "OrderSummary": {
        "description": "A compact order row for lists and the radar.",
        "properties": {
          "created_at": {
            "title": "Created At",
            "type": "string"
          },
          "current_actor_has_rated": {
            "description": "Whether the authenticated actor has rated this order.",
            "title": "Current Actor Has Rated",
            "type": "boolean"
          },
          "delivery_city": {
            "title": "Delivery City",
            "type": "string"
          },
          "delivery_city_id": {
            "format": "uuid",
            "title": "Delivery City Id",
            "type": "string"
          },
          "delivery_date": {
            "title": "Delivery Date",
            "type": "string"
          },
          "description": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Description"
          },
          "id": {
            "title": "Id",
            "type": "string"
          },
          "status": {
            "title": "Status",
            "type": "string"
          }
        },
        "required": [
          "id",
          "status",
          "delivery_city",
          "delivery_city_id",
          "delivery_date",
          "description",
          "created_at",
          "current_actor_has_rated"
        ],
        "title": "OrderSummary",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      },
...
    }
  },
  "info": {
    "title": "Giftly API",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/orders": {
      "get": {
        "description": "List owned/assigned orders newest first, optionally within inclusive delivery dates.\n\nUse Gregorian YYYY-MM-DD bounds. Reset the cursor when dates or status change,\nand retain identical filters on subsequent pages. Responses remain unchanged.",
        "operationId": "list_orders_api_orders_get",
        "parameters": [
          {
            "in": "query",
            "name": "status",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "$ref": "#/components/schemas/OrderStatus"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Status"
            }
          },
          {
            "in": "query",
            "name": "cursor",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "format": "uuid",
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Cursor"
            }
          },
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 20,
              "maximum": 100,
              "minimum": 1,
              "title": "Limit",
              "type": "integer"
            }
          },
          {
            "description": "Inclusive start date (Gregorian YYYY-MM-DD).",
            "in": "query",
            "name": "from_date",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "format": "date",
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Inclusive start date (Gregorian YYYY-MM-DD).",
              "title": "From Date"
            }
          },
          {
            "description": "Inclusive end date (Gregorian YYYY-MM-DD).",
            "in": "query",
            "name": "to_date",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "format": "date",
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Inclusive end date (Gregorian YYYY-MM-DD).",
              "title": "To Date"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OrderListResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "List Orders",
        "tags": [
          "orders"
        ]
      },
      "post": {
        "description": "Create a NEW gift-request order.",
        "operationId": "create_order_api_orders_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateOrderRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OrderDetail"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Create Order",
        "tags": [
          "orders"
        ]
      }
    },
    "/api/orders/available": {
      "get": {
        "description": "List NEW orders in the courier's city.",
        "operationId": "available_orders_api_orders_available_get",
        "parameters": [
          {
            "in": "query",
            "name": "cursor",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "format": "uuid",
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Cursor"
            }
          },
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 20,
              "maximum": 100,
              "minimum": 1,
              "title": "Limit",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OrderListResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Available Orders",
        "tags": [
          "orders"
        ]
      }
    },
    "/api/orders/{order_id}": {
      "get": {
        "description": "Return an order the caller participates in.",
        "operationId": "get_order_api_orders__order_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "order_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Order Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OrderDetail"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Order",
        "tags": [
          "orders"
        ]
      }
    },
    "/api/orders/{order_id}/accept": {
      "post": {
        "description": "Accept a NEW order (Redis lock + FOR UPDATE race).",
        "operationId": "accept_order_api_orders__order_id__accept_post",
        "parameters": [
          {
            "in": "path",
            "name": "order_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Order Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OrderDetail"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Accept Order",
        "tags": [
          "orders"
        ]
      }
    },
    "/api/orders/{order_id}/approve": {
      "post": {
        "description": "Approve a delivered order: complete it and release escrow (customer).",
        "operationId": "approve_order_api_orders__order_id__approve_post",
        "parameters": [
          {
            "in": "path",
            "name": "order_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Order Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OrderDetail"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Approve Order",
        "tags": [
          "orders"
        ]
      }
    },
    "/api/orders/{order_id}/cancel": {
      "post": {
        "description": "Cancel an order before it is in progress.",
        "operationId": "cancel_order_api_orders__order_id__cancel_post",
        "parameters": [
          {
            "in": "path",
            "name": "order_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Order Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CancelOrderRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OrderDetail"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Cancel Order",
        "tags": [
          "orders"
        ]
      }
    },
...
    "/api/orders/{order_id}/dispute": {
      "post": {
        "description": "Open a dispute on an order, freezing escrow (either participant).",
        "operationId": "dispute_order_api_orders__order_id__dispute_post",
        "parameters": [
          {
            "in": "path",
            "name": "order_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Order Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DisputeRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DisputeResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Dispute Order",
        "tags": [
          "orders"
        ]
      }
    },
...
  }
}